% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centre.R
\name{centre}
\alias{centre}
\title{Centre Functional Time Series}
\usage{
centre(x, type)
}
\arguments{
\item{x}{A functional time series object.}

\item{type}{Type of centering: "mean", "var", "median", or "trimmed".}
}
\value{
A centered functional time series.
}
\description{
Centers a functional time series using various methods.
}
\details{
The available centering types are:
\itemize{
  \item "mean": Center by mean function
  \item "var": Center by variance function
  \item "median": Center by median function
  \item "trimmed": Center by trimmed mean function
}
}
\examples{
\dontrun{
# Load example data
data(pm_10_GR)

# Center by mean
centered_data <- centre(pm_10_GR, type = "mean")
}

}
\seealso{
\code{\link{mean.fts}}, \code{\link{median.fts}}
}
