% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dens2lqdNew.R
\name{dens2lqd}
\alias{dens2lqd}
\title{Function for converting densities to log quantile density functions}
\usage{
dens2lqd(
  dens,
  dSup,
  lqdSup = seq(0, 1, length.out = length(dSup)),
  t0 = dSup[1],
  verbose = TRUE
)
}
\arguments{
\item{dens}{density values on dSup - must be strictly positive (otherwise will truncate) and integrate to 1}

\item{dSup}{support (grid) for Density domain}

\item{lqdSup}{support of length M for lqd domain - must begin at 0 and end at 1; (default: seq(0, 1, length(dSup)))}

\item{t0}{value in dSup for which the cdf value c is retained, i.e. c = F(t0) (default: dSup[1])}

\item{verbose}{if FALSE, repress some messages (default: TRUE)}
}
\value{
list with 'lqdSup' a grid on [0,1], 'lqd' the log quantile density on lqdSup, and 'c' the value of the cdf at t0
}
\description{
Function for converting densities to log quantile density functions
}
\examples{
x <- seq(0,2,length.out =512)
y <- rep(0.5,length.out =512)
y.lqd <- dens2lqd( dens=y, dSup = x) # should equate # log(2)
}
\references{
\cite{Functional Data Analysis for Density Functions by Transformation to a Hilbert space, Alexander Petersen and Hans-Georg Mueller, 2016}
}
