% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facf.R
\name{facf}
\alias{facf}
\title{Functional Autocorrelation Function}
\usage{
facf(fun_data, lag_value_range = seq(0, 20, by = 1))
}
\arguments{
\item{fun_data}{A matrix or data frame containing functional time series data.}

\item{lag_value_range}{Vector of lag values to compute autocorrelation for (default: seq(0, 20, by = 1)).}
}
\value{
A vector of autocorrelation values corresponding to the lag values.
}
\description{
Computes the functional autocorrelation function for a functional time series.
}
\details{
The functional autocorrelation function measures the correlation between 
functional observations at different time lags. For lag 0, the value is set to NA.
}
\examples{
\dontrun{
# Load example data
data(pm_10_GR)

# Compute functional autocorrelation
acf_values <- facf(pm_10_GR$y)

# Plot autocorrelation function
plot(seq(0, 20, by = 1), acf_values, type = "l", 
     xlab = "Lag", ylab = "Autocorrelation")
}

}
\seealso{
\code{\link{ftsm}}, \code{\link{forecast.ftsm}}
}
