% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.fts.R
\name{is.fts}
\alias{is.fts}
\title{Check if Object is a Functional Time Series}
\usage{
is.fts(x)
}
\arguments{
\item{x}{An object to be tested.}
}
\value{
Logical. TRUE if the object is a functional time series, FALSE otherwise.
}
\description{
Tests whether an object is a functional time series.
}
\examples{
\dontrun{
# Load example data
data(pm_10_GR)

# Check if it's a functional time series
is.fts(pm_10_GR)
}

}
\seealso{
\code{\link{ftsm}}
}
